release

+++ api/services/ImageService.js
@@ -80,6 +80,12 @@ function createPhotos2(req, photoTags, createProperties) {
 }
 
 function destroyPhotos2(photoIds) {
+  if (!Array.isArray(photoIds)) {
+    photoIds = [];
+  }
+  if (photoIds.length === 0) {
+    return Promise.resolve([]);
+  }
   return Photo.destroy({ id: photoIds })
     .then((destroyedPhotos) => {
       let destroyedCloudinaryPhotos = _.map(destroyedPhotos, (photo) => {
